+++ src/models/espnet.jl
@@ -20,8 +20,8 @@ function espnet(ch_in::Int=3, ch_out::Int=1;   # input/output channels
     # encoder
     e1  = Chain(ConvK3(ch_in, 16; stride=2),
                 BatchNorm(16),
-                ConvPReLU(16),
-                # leakyrelu,
+                # ConvPReLU(16),
+                leakyrelu,
                 Dropout(edrops[1]),
     )
 
@@ -39,15 +39,15 @@ function espnet(ch_in::Int=3, ch_out::Int=1;   # input/output channels
     # decoder
     d2 = Chain(ConvTranspK2(ch_out, ch_out; stride=2),
                BatchNorm(ch_out),
-               ConvPReLU(ch_out),
-            #    leakyrelu,
+            #    ConvPReLU(ch_out),
+               leakyrelu,
                Dropout(ddrops[2]),
     )
     d1 = Chain(ESPBlock1(2*ch_out, ch_out; stride=1), # add=false),
                ConvTranspK2(ch_out, ch_out; stride=2),
                BatchNorm(ch_out),
-               ConvPReLU(ch_out),
-            #    leakyrelu,
+            #    ConvPReLU(ch_out),
+               leakyrelu,
                Dropout(ddrops[1]),
     )
     d0 = Chain(ConvK1(2*ch_out, ch_out),